+++ src/main/java/se/kth/Reduce.java
@@ -1,20 +1,21 @@
 package se.kth;
 
 import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.mapreduce.Reducer;
 
 import java.io.IOException;
 /**
  * Created by salman on 5/2/16.
  */
-public class Reduce extends Reducer<HdfsStatKey, IntWritable, HdfsStatKey, IntWritable> {
+public class Reduce extends Reducer<HdfsStatKey, LongWritable, HdfsStatKey, LongWritable> {
     @Override
-    public void reduce(HdfsStatKey key, Iterable<IntWritable> counts, Context context)
+    public void reduce(HdfsStatKey key, Iterable<LongWritable> counts, Context context)
             throws IOException, InterruptedException {
-      int sum = 0;
-      for (IntWritable count : counts) {
+      long sum = 0;
+      for (LongWritable count : counts) {
         sum += count.get();
       }
-      context.write(key, new IntWritable(sum));
+      context.write(key, new LongWritable(sum));
     }
   }
